requirements.txt
absl-py==2.0.0
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cvzone==1.6.1
cycler==0.12.1
filelock==3.12.4
Flask==3.0.0
Flask-RESTful==0.3.10
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.10.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mediapipe==0.10.7
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.1
numpy==1.26.1
oauthlib==3.2.2
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opt-einsum==3.3.0
optional-django==0.3.0
packaging==23.2
Pillow==10.1.0
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyFirmata==1.1.0
pyparsing==3.1.1
pypng==0.20220715.0
pyserial==3.5
pyTelegramBotAPI==4.14.0
python-dateutil==2.8.2
pytz==2023.3.post1
react==4.3.0
requests==2.31.0
rsa==4.9
six==1.16.0
sounddevice==0.4.6
sympy==1.12
telebot==0.0.5
tensorboard-data-server==0.7.1
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
torch==2.1.0
typing_extensions==4.8.0
urllib3==2.0.5
vboxapi==1.0
Werkzeug==3.0.0
wrapt==1.14.1